Changed to include all the server definitions (use server_includes.h instead of common_includes.h) and #define MYSQL_SERVER 1 before that
Turns out the problem was (YET ANOTHER) conditionally excluded piece
of sql_class.h in a MYSQL_SERVER block that cut out a bunch of definitions
that plugin.h needed.

The solution was simple, but only after a long time investigating.

Good some good notes on further simplifying the include tree, but fixing
this for now...