Replaced use of thd_proc_info() macro with calls to set_proc_info() and get_proc_info() internally.  Introduced functions set_thd_proc_info() and get_thd_proc_info() for external users, i.e., plug-ins.
The set_thd_proc_info() accepted callers info that can be used to
print debug output, but the information was not used. The return
value was changed to void and the old value is not fetched any
more. To be able to get the value of proc_info for external
users, the function get_thd_proc_info() was introduced.

The thd_proc_info() macro called set_thd_proc_info() but almost
never used the return value of set_thd_proc_info() so the macro
was replaced with a call of THD::set_proc_info().