Modified JOIN::flatten_subqueries() to ensure that when a LEFT JOIN is found with semi-joins, we still must finalize the subqueries that were not converted.  Previously, after finding a semi-join and LEFT JOIN, the code returned from the function without cleaning up other joins.
Included is a change to the subselect2.test case which now shows a
different EXPLAIN plan for a LEFT OUTER JOIN with multiple dependent
subqueries.  We should monitor the optimizer team's input on this 
function to see if they come up with an appropriate solution.

But for now, this fixes bug #314911 and all tests pass cleanly.