Read (nearly the whole) index information (key and key parts) out of the proto, not FRM.
Update the table.proto to reflect in server structures and possible values.
Biggest change is IndexPart now references fieldnr and not an embedded field.
This is because we need to set the fieldnr in server and I couldn't find a way to get this info out of protobuf.

Also add an assert in the proto writing part to make sure we handle all possible index flags.

Increasingly we abort() on old FRM file formats as I remove FRM code piece by piece.