fix func_math on x86 32bit.
ROUND() was calling down to my_double_round which triggered a Gcc 4.3 bug where isinf()
could be inlined to use the 387 FPU register to do the check on the result of the multiplication.
This would do the the check on the 80bit result, not the truncated 64bit result.

See http://gcc.gnu.org/bugzilla/show_bug.cgi?id=39228
See http://bugs.mysql.com/bug.php?id=42965