Fix key_cache test for 32bit.
The key cache block structure has pointers in it.

On 32 and 64 bit, these are different sizes.

There's even a comment in the test to indicate this.

The test file should replace all possible numbers with KEY_BLOCKS_UNUSED, but this was messed up in fixing the
test case for Drizzle. This cset fixes that.

Currently tested on x86 and x86-64. Pretty sure (99.9%) that this will cover all platforms involved.

Our numbers are different from current MySQL 6 for some random reason that I don't really care about.
Number of blocks * block_size roughly equals key cache size - so i'm happy.