This commit does two things:
Removes lock on serial event log and uses a file offset to pwrite() to the
log file instead of locking it.

Adds an --serial-event-log-enable option for the serial event log plugin and
adds an enable() and disable() method controlling an atomic bool which allows
online disabling of the event log.

Also, adds an optimization to the registration process of replicators and
appliers for TransactionServices which evaluates active plugins on attach and
detach of a replicator and applier.  This enables a quick atomic bool comparison
in the write methods of TransactionServices, essentially making it a NOOP when
no active appliers or replicators are registered with the TransactionServices
component.