Remove keep_files_on_create
Only applicable to MyISAM, and should be down in the engine layer as an error if files trying to create already exist.

Also, if you're wanting to re-create the FRM/DFE, then just replace the data files!
Or at least have it as a CREATE TABLE option, and not some server variable that changes behaviour of
such things.