Beginnings of reworking the mysql_execute_command() method.
Essentially, we want to remove the giant switch statement within that
function. This commit sets the groundwork for that. We added a new base
class named SqlCommand that every command we want to implement inherits
from. It has an execute method that commands can override. Thus, instead of
a large switch statement, we will just have command->execute().