Fix mtr2 timeouts/spurious test failures (bug lp:397912)
This was not mtr2 related at all, but in fact e2fsprogs related.
Namely, libuuid.

or rather, uuidd.

libuuid will fork off uuidd which is a setuid binary.

If we needed to restart the server after using a UUID function, the uuidd
process would still exist and hold onto the drizzled server socket.

This is because the server socket for drizzled wasn't CLOEXEC.

This patch sets FD_CLOEXEC on the server socket so that it's closed on exec,
and the setuid uuidd won't hold onto our server socket.

foiled by unix.

never assume your program doesn't fork and exec a setuid process :/