Adds support for BLOB fields in the replication stream. Note that this does not yet support extremely large BLOB data.  Extremely large BLOB data will be written to the transaction log in specialized transaction log entries for that.  In this patch, we simply allow the main Transaction message and the reader/writer programs to work with binary data properly.
Adds a simple test for BLOB fields and the transaction log.
The problem was figuring out how to correctly replace embedded NUL
characters in the output of the transaction reader with \\0 so that
mysqltest.cc wouldn't puke when cat'ing the result of the reading
of the transaction log when BLOB data was contained.