1) Reduced the size of VARIABLE_VALUE. 2) Shifted tests around to be in their suite.
There is still a bug in VARIABLES, but the bug in not in DD, but somewhere
up above (and as Padraig pointed out in IM, the values were messed up
before).

I noticed while debugging that the values for SHOW_TYPE are not always
right. I am going to go down the rabit hole a bit and take a look at what is
up.

Doing an ORDER BY on VARIABLES creates issues. I created a test for that,
but have it commented out for the moment.