basic UPDATE row for Embedded InnoDB Engine.
Factor out read_row_from_innodb to *not* clear the tuple, so that entering any function we have the tuple filled out for the previous record. To perofrma na update, we then move the cursor back a row (as cursor points to *NEXT* row) and we can then use the tuple (which has not been cleared) to perform the update.

This means that the tuple is now cleared on the *next* Cursor call. Calls to read_row_from_innodb() have been updated accordingly.

We also now have a write_row_to_innodb_tuple() method that takes an array of Field objects and uses them to write out the record to the supplied innodb tuple. This can be used in write_row or update_row.