Two little problems were preventing the plugin from being loaded:
#include <exception> needed to be placed in the handler source
file to properly refer to std::exception.

Had two variables declared extern in rabbitmq_log.cc but those
variables were not defined elsewhere.  Simple fix...change extern
to static to have the variables defined in that file locally.

Adds a test suite for the rabbitmq module with a simple test case
for variables.