This removes the caller for show create table to be a derived table.
It also highlights that SHOW CREATE TABLE was incorrectly reporting whether
a table existed or not (and tableprotester plugin was making use of this
incorrectly).

Confirmation of a table creation needs to happen before, not after the table
has been created.

There is something up in the jp_enum test, I will fix this in another patch
in this series.