Merge Revision revid:svn-v4:16c675df-0fcb-4bc9-8058-dcc011a37293:branches/zip:6740 from MySQL InnoDB
Original revid:svn-v4:16c675df-0fcb-4bc9-8058-dcc011a37293:branches/zip:6740

Original Authors: sunny
Original commit message:
branches/zip: Don't print the entire lock bit set if the block was not
found in the buffer pool. Only print the bits that are set and that
information is in the lock and not in the block.

See rb://256 approved by Marko.