Merge Revision revid:magne.mahre@sun.com-20100617205135-74yw3t2pxs4i23px from MySQL InnoDB
The only things we're taking from this are:
- file_per_table as default
- file_format defaulting to Barracuda


Original revid:magne.mahre@sun.com-20100617205135-74yw3t2pxs4i23px

Original Authors: Magne Mahre <magne.mahre@sun.com>
Original commit message:
WL#5349 Change default storage engine to InnoDB

The default storage engine is changed from MyISAM to
InnoDB, in all builds except for the embedded server.

In addition, the following system variables are 
changed:

  * innodb_file_per_table is enabled
  * innodb_strict_mode is enabled
  * innodb_file_format_name_update is changed
    to 'Barracuda'

The test suite is changed so that tests that do not
explicitly include the have_innodb.inc are run with
--default-storage-engine=MyISAM.  This is to ease the
transition, so that most regression tests are run
with the same engine as before.

Some tests are disabled for the embedded server
regression test, as the output of certain statements
will be different that for the regular server
(i.e SELECT @@default_storage_engine).  This is to
ease transition.