HailDB cursor not reset to clustered index.
Initial symptom was Sysbench failing with HailDB.

Investigation showed that we weren't resetting the cursor we track in the HailDBCursor object (to help avoid lots of malloc/free cycles we keep a cursor around) back to the clustered index after an operation on a secondary index (e.g. index scan). This mean we were attempting the UPDATE on the wrong index, ending up with a DB_ERROR from ib_cursor_update_row(), which translates to -1 (hence "error -1 from engine").