Allow/ignore certain semantic grouping errors that may occur during transformer validation of queries with MySQL systems. The goal is to easily be able to use sql_mode=ONLY_FULL_GROUP_BY with MySQL when the Transformer validator is used. Without this change, all errors caused by this sql-mode would cause the test to abort with STATUS_ENVIRONMENT_FAILURE.
This change allows the following MySQL errors to be produced by Transformers, hence allowing the test to continue:
 1004: ER_NON_GROUPING_FIELD_USED
 1055: ER_WRONG_FIELD_WITH_GROUP
 1056: ER_WRONG_GROUP_FIELD
 1140: ER_MIX_OF_GROUP_FUNC_AND_FIELDS

In normal mode each of these errors will be mentioned in the test output the first time they occur, and further errors will be suppressed.
In debug mode each occurrence of all such errors will be mentioned along with the offending queries.

Other types of semantic or syntactic errors, and executors that are not of the type DB_MYSQL, will still return STATUS_ENVIRONMENT_FAILURE.