New ErrorLogAlarm Reporter, and new status STATUS_ALARM (109). The ErrorLogAlarm Reporter is a periodic reporter which scans the server's error log for a specific regex pattern and returns STATUS_ALARM if this pattern is found. The default pattern is: ^ERROR
Using this Reporter will then allow users to be alerted when
an error occurs in the server error log, for example if
an init file contains a typo.

The pattern can be customized to for example look for 
specific valgrind warnings or other messages instead.
Customization is currently only supported by editing the 
source code.

STATUS_ALARM was added because the test should be aborted if
the pattern is found, and there was no other critical status 
already available that seemed suitable for this situation.