Use v8::Locker to allow multiple threads to use v8 engine.
Note that making all threads that need to use JavaScript lock
on a single mutex is going to be problematic. Newer versions
of libv8 provide a class Isolate, through which each thread can
have their own, isolated, instance of a v8 engine and use it
without locking. However, current version in Ubuntu 11.04 is
v8 2.5.9.9 and doesn't yet provide this class. Future work
on this code should add a run-time check about the v8 version
and use Isolate if available and Locker otherwise.