http://hostname:8086/json now can do a single row get. Result is returned in json and also has to be stored in json. There's no insert yet, you'd have to use SQL for that.
Also updated the web gui to showcase this new api.

Also fixes an off by one bug in drizzled/plugin/client/cached.h
Sorry about not making that a separate bug, but I didn't know of any 
other code that would use that class, so I had to do it within this code.