Can now insert and update records with HTTP POST + JSON. This was very handy to do with REPLACE INTO as the philosophy is the same as JSON databases do it.
Made the web ui support selecting GET/POST.

Column names of the table are not hard coded. You can create a table
with any number of TEXT columns with any names, as long as there is an
_id auto_increment primary key. (GET supports this too now.)

Had to extend the result set in Execute API to support getting column
metadata. (Previously the only meta data supported was number of columns.)