Support for signing packages. Packages are signed if $DEBEMAIL and $DEBFULLNAME match a key in your GPG keyring.
make deb-source
Just prepare the .changes .dsc .orig.tar.gz and .debian.tar.gz files
but not binary packages. This is what you will upload to Launchpad
PPA. You must provide $DEBEMAIL and $DEBFULLNAME with a valid
gpg key.

Changed to use 7.1.31~rc version tags instead of 7.1.31-rc. Using a dash
is impossible in rpms and while it is possible for debian, it has side
effects and is frowned upon.

Appended ~oneiric1 to deb version so that same repo can contain debs
for multiple distro versions.